source/rest_api/requirements.txt
chromadb==0.5.0
fastapi==0.111.0
langchain==0.1.17
langchain_core==0.1.50
langchain_ibm==0.1.4
python-dotenv==1.0.1
uvicorn==0.29.0
youtube_transcript_api==0.6.1
